+++ EarthsTimeline/Controllers/CommentsController.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using EarthsTimeline.Models;
 
@@ -18,15 +15,26 @@ namespace EarthsTimeline.Controllers
             _context = context;
         }
 
+        private bool LoggedIn()
+        {
+            if (Request.Cookies.ContainsKey("AntiForge") &&
+                Request.Cookies["AntiForge"] == "UUDDLRLRBABAS")
+                return true;
+            else
+                return false;
+        }
+
         // GET: Comments
         public async Task<IActionResult> Index()
         {
+            if (!LoggedIn()) return Redirect("~/login");
             return View(await _context.Comment.ToListAsync());
         }
 
         // GET: Comments/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!LoggedIn()) return Redirect("~/login");
             if (id == null)
             {
                 return NotFound();
@@ -42,31 +50,10 @@ namespace EarthsTimeline.Controllers
             return View(comment);
         }
 
-        // GET: Comments/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Comments/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Approved,PostId,Date,AuthorId,AuthorName,Content")] Comment comment)
-        {
-            if (ModelState.IsValid)
-            {
-                _context.Add(comment);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(comment);
-        }
-
         // GET: Comments/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!LoggedIn()) return Redirect("~/login");
             if (id == null)
             {
                 return NotFound();
@@ -87,6 +74,7 @@ namespace EarthsTimeline.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Approved,PostId,Date,AuthorId,AuthorName,Content")] Comment comment)
         {
+            if (!LoggedIn()) return Redirect("~/login");
             if (id != comment.Id)
             {
                 return NotFound();
@@ -118,6 +106,7 @@ namespace EarthsTimeline.Controllers
         // GET: Comments/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!LoggedIn()) return Redirect("~/login");
             if (id == null)
             {
                 return NotFound();
@@ -138,6 +127,7 @@ namespace EarthsTimeline.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!LoggedIn()) return Redirect("~/login");
             var comment = await _context.Comment.FindAsync(id);
             _context.Comment.Remove(comment);
             await _context.SaveChangesAsync();
